+++ CHANGELOG.md
@@ -2203,4 +2203,4 @@ Alternately, the _Daily Diary of my Programming Dreams_.
 
 * **Angular-Cli**
 
-    1. Added [fiddles/angular2-cli/fiddle-0030-EntAgGridChartRangeSelect](fiddles/angular2-cli/fiddle-0030-EntAgGridChartRangeSelect)
+    1. Added [fiddles/angular2-cli/fiddle-0030-EntAgGridChartRangeSelect](fiddles/angular2-cli/fiddle-0030-EntAgGridChartRangeSelect)* Added [fiddles/angular2-cli/fiddle-0031-NgxBootstrapBasics](fiddles/angular2-cli/fiddle-0031-NgxBootstrapBasics)